feat(ui): extract JsonDropZone component
